Refactors to base projects
